Updating information
A few spots of CIP1 pointed towards the now offline (old) Counterparty Forums. 

I changed the links to the new forum links that denote the correct area the CIP1 wanted to point new developers and users to.

This includes:
The Counterparty Forums (fixed link)
The CIP Discussions github
The Counterparty Developer Telegram

I also changed the last link which denoted sending Devon Weller an email when JPJA is now the CIP Editor.